+++ main/OrbisGL/Shaders/VertexOffsetTexture.glsl
@@ -5,7 +5,7 @@ attribute vec2 uv;
 
 uniform vec3 Offset;
  
-varying lowp vec2 UV;
+varying highp vec2 UV;
 
 void main(void) {
     gl_Position = vec4(Position + Offset, 1.0);
